# cmd/docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Review:
    properties:
      animeId:
        description: ID of the anime being reviewed
        example: 1
        type: integer
      content:
        description: Content of the review
        example: This anime is amazing!
        maxLength: 500
        type: string
      created_at:
        description: Timestamp when the review was created
        example: "2025-02-20T19:27:00Z"
        type: string
      id:
        description: Unique identifier for the review
        example: 1
        type: integer
      rating:
        description: Rating given in the review (0-10)
        example: 9
        maximum: 10
        minimum: 0
        type: integer
      updated_at:
        description: Timestamp when the review was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      userId:
        description: ID of the user who created the review
        example: 1
        type: integer
    required:
    - content
    - rating
    type: object
...
  models.User:
    properties:
      created_at:
        description: Timestamp when the user was created
        example: "2025-02-20T19:27:00Z"
        type: string
      email:
        description: Email address of the user (omitted unless necessary)
        example: john@example.com
        maxLength: 100
        minLength: 5
        type: string
      id:
        description: Unique identifier for the user
        example: 1
        type: integer
      is_admin:
        description: Indicates if the user has admin privileges
        example: false
        type: boolean
      reviews:
        description: List of reviews created by the user (omitted unless necessary)
        items:
          $ref: '#/definitions/models.Review'
        type: array
      updated_at:
        description: Timestamp when the user was last updated
        example: "2025-02-20T19:27:00Z"
        type: string
      username:
        description: Unique username for the user
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
...
  models.UserCredentials:
    properties:
      password:
        description: Password for authentication
        example: password123
        maxLength: 100
        minLength: 5
        type: string
      username:
        description: Username for authentication
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@myanimeapi.com
    name: API Support
  description: This is a sample API for managing anime and reviews.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: MyAnimeAPI
  version: "1.0"
paths:
...
  /v1/auth/authenticate:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.UserCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: Returns a JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not found or invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to generate token
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Authenticate a user
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided data
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User with this username or email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to hash password or create user
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - "":
        - ""
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Use the format "Bearer <JWT_TOKEN>". Example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"